Add simple terminal ui version
This terminal version currently utilizes the linux/keyboard.h header for
access to keyboard state. This is not ideal, and alternatively we could
utilize X11 for key events if we are running an x-server.

This has been partially looked at, the main stoppage point currently is
retrieving the current X11 window from the parent process Id. Or, the
larger problem, how do we find the terminal window id that the program
was run in easily?

+++ src/mpe/block.cpp
@@ -75,7 +75,7 @@ block::block(const block_type id, const rotation_type r)
 
 bool block::collision(const field &field)
 {
-    for (int i = 0; i < data.size(); ++i) {
+    for (size_t i = 0; i < data.size(); ++i) {
         const int xoffset =  x + data[i].x;
         const int yoffset =  y + data[i].y;
 
@@ -215,7 +215,7 @@ void block::hard_drop(const field &field)
 
 bool block::at(const int xl, const int yl) const
 {
-    for (int i = 0; i < data.size(); ++i) {
+    for (size_t i = 0; i < data.size(); ++i) {
         const int bx =  x + data[i].x;
         const int by =  y + data[i].y;
